Rename report_generator to reporter for consistency
